start using the 'scan_and_wait' from walker
now the post call to /scan will schedule the walker to wait till all
the thumber jobs in the queue are finished.

this helps in calling scan independent of status of compress jobs. so
the client can call scan and it is guaranteed  that all uploaded images
will be available for download after compression.

but if the any images are upload while the scanner is running, these
images will not be available for download by this version of scanner and
will require a  new version of scanner to be queued but that too after
the existing one is finished, this is a rare race condition